+++ Arcadia/Views/DiscoverGameListView.swift
@@ -1,0 +1,18 @@
+//
+//  DiscoverGameListView.swift
+//  Arcadia
+//
+//  Created by Davide Andreoli on 27/08/24.
+//
+
+import SwiftUI
+
+struct DiscoverGameListView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    DiscoverGameListView()
+}