+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## Features
 - Added search bar in game list
+- Added game validity check: if you somehow manage to load a game that is not compatible with the system you are loading it in, you get an alert and the game view is dismissed
+- Added the possibility to redownload the default cover image
 
 ## Improvements
 - Improved recommendation results retrieval speed
+- Improved L and R buttons shape
 
+## Bug fixes
+- Controller shoulder buttons are correctly managed
 
 # 1.2